+++ chef/cookbook/cookbook_test.go
@@ -5,16 +5,16 @@ import (
 	"os"
 	"testing"
 
-	. "github.com/mlafeldt/chef-runner/cookbook"
+	. "github.com/mlafeldt/chef-runner/chef/cookbook"
 	"github.com/mlafeldt/chef-runner/util"
 	"github.com/stretchr/testify/assert"
 )
 
 func TestNewCookbook(t *testing.T) {
-	cb, err := NewCookbook("../testdata")
+	cb, err := NewCookbook("../../testdata")
 	assert.NoError(t, err)
 	if assert.NotNil(t, cb) {
-		assert.Equal(t, "../testdata", cb.Path)
+		assert.Equal(t, "../../testdata", cb.Path)
 		assert.Equal(t, "practicingruby", cb.Name)
 		assert.Equal(t, "1.3.1", cb.Version)
 	}
@@ -36,12 +36,12 @@ func TestString(t *testing.T) {
 }
 
 func TestFiles(t *testing.T) {
-	cb, _ := NewCookbook("../testdata")
+	cb, _ := NewCookbook("../../testdata")
 	expect := []string{
-		"../testdata/README.md",
-		"../testdata/metadata.rb",
-		"../testdata/attributes",
-		"../testdata/recipes",
+		"../../testdata/README.md",
+		"../../testdata/metadata.rb",
+		"../../testdata/attributes",
+		"../../testdata/recipes",
 	}
 	assert.Equal(t, expect, cb.Files())
 }